requirements.txt
torch
transformers
faiss-cpu
PyPDF2
groq
